Refactor RustGenerator into another module
Part of https://github.com/google/sedpack/issues/200
